Create Satellite Position
I made this code as part of my Texas A&M AggieSat (Satellite) Club.

The code is the beginning of a process of transforming mathlab
code that was developed to acquire the position of the satellite 
in orbit into C++ code. This is just a basic algorithm done to more
basic computations.